requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_tflow_select=2.1.0=gpu
absl-py=0.6.1=py36_0
alabaster=0.7.12=py36_0
asn1crypto=0.24.0=py36_0
astor=0.7.1=py36_0
astroid=2.2.5=py36_0
attrs=19.1.0=py_0
babel=2.6.0=py36_0
backcall=0.1.0=py36_0
blas=1.0=mkl
bleach=3.1.0=py36_0
c-ares=1.15.0=h7b6447c_1
ca-certificates=2019.1.23=0
cairo=1.14.12=h8948797_3
certifi=2019.3.9=py36_0
cffi=1.12.2=py36h2e261b9_1
chardet=3.0.4=py36_1
click=7.0=pypi_0
cloudpickle=0.6.1=py36_0
cryptography=2.6.1=py36h1ba5d50_0
cudatoolkit=9.2=0
cudnn=7.2.1=cuda9.2_0
cupti=9.2.148=0
cycler=0.10.0=py36_0
cytoolz=0.9.0.1=py36h14c3975_1
dask-core=1.0.0=py36_0
dbus=1.13.2=h714fa37_1
decorator=4.3.0=py36_0
defusedxml=0.5.0=py36_1
docutils=0.14=py36_0
entrypoints=0.3=py36_0
expat=2.2.6=he6710b0_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
fribidi=1.0.5=h7b6447c_0
gast=0.2.0=py36_0
glib=2.56.2=hd408876_0
gmp=6.1.2=h6c8ec71_1
graphite2=1.3.12=h23475e2_2
graphviz=2.40.1=h21bd128_2
grpcio=1.16.1=py36hf8bcb03_1
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.8.0=py36h989c5e5_3
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=h9c2bf20_1
idna=2.8=py36_0
imageio=2.4.1=py36_0
imagesize=1.1.0=py36_0
intel-openmp=2019.1=144
ipykernel=5.1.0=py36h39e3cac_0
ipython=7.2.0=py36h39e3cac_0
ipython_genutils=0.2.0=py36_0
isort=4.3.15=py36_0
jedi=0.13.2=py36_0
jeepney=0.4=py36_0
jinja2=2.10=py36_0
jpeg=9b=h024ee3a_2
jsonschema=3.0.1=py36_0
jupyter_client=5.2.4=py36_0
jupyter_core=4.4.0=py36_0
keras=2.2.4=0
keras-applications=1.0.6=py36_0
keras-base=2.2.4=py36_0
keras-preprocessing=1.0.5=py36_0
keyring=18.0.0=py36_0
kiwisolver=1.0.1=py36hf484d3e_0
lazy-object-proxy=1.3.1=py36h14c3975_2
libedit=3.1.20170329=h6b74fdf_2
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.35=hbc83047_0
libprotobuf=3.6.1=hd408876_0
libsodium=1.0.16=h1bed415_0
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.9=he85c1e1_2
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.8=h26e45fe_1
markdown=3.0.1=py36_0
markupsafe=1.1.1=py36h7b6447c_0
matplotlib=3.0.2=py36h5429711_0
mccabe=0.6.1=py36_1
mistune=0.8.4=py36h7b6447c_0
mkl=2019.1=144
mkl_fft=1.0.6=py36hd81dba3_0
mkl_random=1.0.2=py36hd81dba3_0
nbconvert=5.4.1=py_2
nbformat=4.4.0=py36_0
ncurses=6.1=he6710b0_1
networkx=2.2=py36_1
numpy=1.15.4=py36h7e9f1db_0
numpy-base=1.15.4=py36hde5b4d6_0
numpydoc=0.8.0=py36_0
olefile=0.46=py36_0
opencv-contrib-python=3.4.4.19=pypi_0
openssl=1.1.1b=h7b6447c_1
packaging=19.0=py36_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py36_1
pango=1.42.4=h049681c_0
parso=0.3.1=py36_0
pcre=8.42=h439df22_0
pexpect=4.6.0=py36_0
pickleshare=0.7.5=py36_0
pillow=5.3.0=py36h34e0f95_0
pip=18.1=py36_0
pixman=0.34.0=hceecf20_3
prompt_toolkit=2.0.7=py36_0
protobuf=3.6.1=py36he6710b0_0
psutil=5.6.1=py36h7b6447c_0
ptyprocess=0.6.0=py36_0
pycodestyle=2.5.0=py36_0
pycparser=2.19=py36_0
pydot=1.4.1=pypi_0
pyflakes=2.1.1=py36_0
pygments=2.3.1=py36_0
pylint=2.3.1=py36_0
pyopenssl=19.0.0=py36_0
pyparsing=2.3.0=py36_0
pyqt=5.9.2=py36h05f1152_2
pyrsistent=0.14.11=py36h7b6447c_0
pysocks=1.6.8=py36_0
python=3.6.7=h0371630_0
python-dateutil=2.7.5=py36_0
python-graphviz=0.10.1=pypi_0
pytz=2018.7=py36_0
pywavelets=1.0.1=py36hdd07704_0
pyyaml=3.13=py36h14c3975_0
pyzmq=17.1.2=py36h14c3975_0
qt=5.9.7=h5867ecd_1
qtawesome=0.5.7=py_0
qtconsole=4.4.3=py36_0
qtpy=1.7.0=py_0
readline=7.0=h7b6447c_5
requests=2.21.0=py36_0
rope=0.12.0=py36_0
scikit-image=0.14.1=py36he6710b0_0
scipy=1.1.0=py36h7c811a0_2
secretstorage=3.1.1=py36_0
setuptools=40.6.3=py36_0
sip=4.19.8=py36hf484d3e_0
six=1.12.0=py36_0
snowballstemmer=1.2.1=py36_0
sphinx=1.8.5=py36_0
sphinxcontrib=1.0=py36_1
sphinxcontrib-websupport=1.1.0=py36_1
spyder=3.3.3=py36_0
spyder-kernels=0.4.2=py36_0
sqlite=3.26.0=h7b6447c_0
tensorboard=1.12.1=py36he6710b0_0
tensorflow=1.12.0=gpu_py36he74679b_0
tensorflow-base=1.12.0=gpu_py36had579c0_0
tensorflow-gpu=1.12.0=h0d30ee6_0
termcolor=1.1.0=py36_1
testpath=0.4.2=py36_0
tk=8.6.8=hbc83047_0
toolz=0.9.0=py36_0
tornado=5.1.1=py36h7b6447c_0
traitlets=4.3.2=py36_0
typed-ast=1.3.1=py36h7b6447c_0
urllib3=1.24.1=py36_0
wcwidth=0.1.7=py36_0
webencodings=0.5.1=py36_1
werkzeug=0.14.1=py36_0
wheel=0.32.3=py36_0
wrapt=1.11.1=py36h7b6447c_0
wurlitzer=1.0.2=py36_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
zeromq=4.2.5=hf484d3e_1
zlib=1.2.11=h7b6447c_3
